Merged PR 1158: Add logical ops: and/or/xor/less/greater/equal, plus new OpTester class
The logical ops were the first ones to mix types, so the existing SimpleFloatTest wasn't enough. Plus I thought I could simplify things further.

In OpTester, it generates the input & output defs automatically based on the inputs & outputs set. See "OpTester test("Equal");" in the change for an example. It should be self explanatory.

I didn't convert all of the existing tests over yet, as this is an incremental change. I will convert and remove the old code in a future change.

Related work items: #177