Merged PR 1206: Add support for getting model metadata.
Add support for getting model metadata. Will add test cases once the initial review looks good.

Note that this is slightly different from the LotusRT versions of GetInputs/GetOutputs in that the NodeArg is not directly exposed to the clients of the session class.

Also, notice the return by pair<>. Did this to achieve 2 things: (1) ensure that an appropriate error code is returned if the client didn't load the model to begin with (2) avoid copying of the metadata. The LotusRT version didn't check if the model was loaded or not when these functions were called.

Related work items: #178