Merged PR 1289: Miscellaneous improvements to OpTester::Run
Update OpTester::Run to catch and output exceptions so it's easier to see error output if you're not expecting a test to throw.
Update OpTester::Run to check status from graph->Resolve().
Adjust ordering of operations in OpTester::Run so that the attributes are added to the Node prior to Resolve being called.
Fix some inconsistent names that got missed.