Merged PR 1418: Add test data dirs from the ONNX submodule to list used by onnx_test_runner.
Add test data dirs from the ONNX submodule to list used by onnx_test_runner.

Have made this optional for now based on the existing --enable_onnx_tests flag given a number of tests fail and there's a lot of output. When the tests are passing it should become the default, and failure should return an error so it is considered a build break. Or the cmake setup should add the onnx tests instead of this script.

Related work items: #437