Merged PR 1569: Windows: Memory leaks in lotus lib
I'm 99% sure that the bug was that '::google::protobuf::ShutdownProtobufLibrary()' wasn't being called in that case, as we had our individual test projects calling it. It makes more sense to have the Environment call this on destruction so that it happens automatically at a known time.

Another possibility is to have a static in environment.cc that does the same thing at a slightly later point in time.

Related work items: #560