Merged PR 1723: General Broadcasting Support for elementwise ops
I haven't written out any detailed comments in the code as we really wanted to get this checked in to unblock other tests.

I'll write out comments and add more tests soon, but the general idea of the code is to analyze the input shapes and then figure out the largest block it can do per pass, being either vector*scalar, scalar*vector, or vector*vector (the three cases seen in each op). For tensors with the same shape, it's always a single iteration, so there is no advantage to bypassing the broadcasting code.

Related work items: #622, #658, #682