Merged PR 2042: Fix a 32-bit bug in the broadcasting code
A negative size_t was being converted into an int64_t which resulted in the wrong value.

The problem is a negative unsigned value is still unsigned, and then it was stored in an int64_t. So the result was a large positive value instead of a small negative one. Fix is to use the right type.

Related work items: #886