Merged PR 2096: Improve lotus_exec error reporting
Be consistent about outputting errors on cerr, and messages on cout. Most places use cerr for exception messages with the exception of main.cc which was backwards.
Include the actual error message from an MLValue mismatch to the reason for the mismatch is in the output.

NOTE: There's a separate winmltools PR to update fn_model_conversion.py to read the status from stdout (matches changes to main.cc here) that will be checked in at the same time.