Merged PR 2186: Restore ability of build.py to install required APT packages on WSL if required.
If you just want to do the occasional local linux build WSL is better than Docker as it doesn't require reserving RAM for a VM to run docker in (and docker doesn't support dynamic memory so the VM memory usage is fixed). Restore the ability for build.py to work nicely on WSL if there are missing APT repositories/packages required.

Also be graceful about requiring the script to be run using sudo. Don't ask if not needed. Tell user what to do if it is.