Merged PR 2351: Change USE_MLAS #ifdef to check _M_AMD64 instead of _AMD64_
Razzle automatically defines _AMD64_ for all code, but this is not defined until including <windows.h> in the outside world, so the VSTS based build isn't matching the Razzle build. To make this consistent everywhere, cange the #ifdef check to use _M_AMD64 instead. This time actually verified that resnet50_fp16 is fast for both Razzle and mscodehub build.

Related work items: #1130