Merged PR 2390: Return reference from Model::MainGraph and update places that require a Graph instance to use a reference in the API instead of a pointer.
As a valid Model should always have a graph, return a reference from Model::MainGraph and enforce the invariant that there is a graph. The enforcing of this invariant is the only significant change.

Trivial flow on changes to take Graph as reference instead of pointer so APIs indicate clearly a graph instance is required (was implicit previously). https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#Rf-ptr-ref

Graph is not copyable so there is no risk of accidental copies of the return value from Model::MainGraph.

Started making changes as part of control flow work but splitting them out now so the PR is much smaller.

Related work items: #131