Merged PR 2434: Support the creation of a shared library for Linux + add simple and custom op inference test using the created shared lib.
Note that I've enabled this creation through a flag --build_shared_lib to build.py. When this flag is used, jemalloc is disabled. This is because jemalloc cannot be used as-is to create the shared lib (see https://aiinfra.visualstudio.com/Lotus/_workitems/edit/1234).

We should use --build_shared_lib by default in the Linux CI build. Will configure that.