Merged PR 2444: Add an empty cc file to make cmake happy when building a shared lib from archives only.
Add an empty cc file to make cmake happy when building a shared lib from archives only. cmake 3.10.2 seems to be ok without this change it just prints a warning, but looks like the cmake on the build machine is not forgiving.

For reference see here https://github.com/apache/incubator-mxnet/pull/10711