Merged PR 2497: Allow some partners to contribute ops in Lotus (see rationale etc below)
Some of our partners need new ops to be implemented by Lotus. Today this can be achieved by writing a custom op and registering it with Lotus either by calling RegisterCustomRegistry or loading it inside Lotus as a DLL via LoadCustomOps. However for partners like AutoML this doesn't work very well as they only generate a model based on user data/requirements. Obviously the model doesn't include the custom op implementations. Hence the users cannot just run the model on any AML (azure ml) image. In order to support their users AutoML can potentially supply a SDK to it's users or make a pre-built image available in AML that takes care of handling the op implementation registrations. However this has 2 problems: users are not free to run the model on any image and second AutoML doesn't want to maintain this SDK. As we're beginning to increase ONNX/Lotus adoption I think we should make it as friction-less as possible. Additionally some of these ops are in the process of getting adding to ONNX but their release timelines doesn't match that of ONNX. This PR alleviates these concerns by allowing some of our partners to add new ops in a new "contrib" folder and adds an example op for demonstration.

Some common questions and their answers:
1. Who will support these ops?
A. Lotus. So far for AutoML we've 4 ops and out of which 3 will be submitted to ONNX any way at some point of time. So that leaves only one op for support.
2. Will these be available for all Lotus users?
A. Yes. Hence we should limit these to only standard ops that cannot be implemented using RegisterCustomRegistry or LoadCustomOps and which do not make it to Lotus.
3. What if there is a bug in the op and Lotus release timelines don't match theirs?
A. Unfortunately they'll have to wait for Lotus release. We should make this clear to them.

Related work items: #1256