Merged PR 2518: Prototype for integrate tvm to Lotus
This prototype is targeting enable developers to fuse parts of Lotus graph and run it with a tvm kernel. Two scenarios might need to consider: 1. Empower the speech team to integrate their LSTM optimization to Lotus. 2. Use this workflow to generate FPGA code and run it in Lotus. The prototype contain an example about how to fuse continued Add op together and run that part on tvm.