Merged PR 2907: patch mkldnn linux/gcc build. it should not build with -march=native -mtune=native
when built with gcc, mkldnn is building with -march=native -mtune=native
this means if it is built on machine with avx (i.e. our build machines)
lotus will crash on machine that does not support avx instructions.
perf impact should be negligible (mkldnn primitives will use avx instructions if the target machine supports it)