Merged PR 2981: Fix model test build by handling unused initializer that is in the graph inputs better.
We need to maintain the graph input order, so an unused initializer should remain in them. Previously it was deleted, which changes the input order.

Update the session state initializer to create entries in the MLValue to idx map for all graph inputs as the allocation planner iterates them and requires all inputs to have an entry.

We could alternatively change the allocation planner to allow things to be missing an entry, but that seems less robust than adding an entry for something that is never used.

Related work items: #1687