Merged PR 3125: Optimize SGEMM(TransB=CblasTrans) for ARM/x86
Optimize the case of SGEMM where the B matrix is transposed. x64 already had a fast path that shipped in RS5, this PR adds a path using NEON or SSE2 intrinsics for ARM/ARM64/x86. I also verified that these changes build with Linaro toolchain for ARM/ARM64. The improvement is noticeable in some smaller LSTM models in another change I'm cooking up. The changes were tested in the standalone unit test for the various architectures.